+++ RELEASENOTES.md
@@ -35,6 +35,7 @@
 - Fix to match boom behavior to parsing dehacked integers with failures defaulting to zero.
 - Fix SBARDEF crash when children element was explicity set to null.
 - Fix dehacked dropped ammo types only giving half ammo.
+- Fix "Compact HUD" (Tate/Portrait Window) behavior for SBARDEF
 
 ## Misc:
 - Refactor of old Status Bar renderer to data-driven SBARDEF format.